Revert "Bh dev: fixed several bugs Arm_3 now essentially working"
