+++ ikSolver.py
@@ -128,8 +128,8 @@ SolverDebug = False
 ###add in new nodes:assigner and rank node#############
 asgn = assigner()
 asgn.Name = "Assigner"
-rankNode = rank()
-rankNode.Name = "Rank Node"
+rk = rank()
+rk.Name = "Rank Node"
 #######################################################
 tanID = tan_id()
 tanID.Name = 'Tangent ID'
@@ -146,11 +146,11 @@ tanSol.BHdebug =  True #LeafDebug
 
 algID = algebra_id()
 algID.Name = "Algebra ID"
-algID.BHdebug = True #LeafDebug
+algID.BHdebug = LeafDebug
 
 algSolver = algebra_solve()
 algSolver.Name = "Algebra Solver"
-algSolver.BHdebug = True
+algSolver.BHdebug = LeafDebug
 
 algSol = b3.Sequence([algID, algSolver])
 algSol.Name = "Algebra ID and Solve"
@@ -224,12 +224,14 @@ compDetect.BHdebug = True
 #   Higher level BT nodes here
 #
 
-sc_tan = b3.Sequence([b3.OrNode([tanSol, scSol]), rankNode])
+sc_tan = b3.Sequence([b3.OrNode([tanSol, scSol]), rk])
 
 
 # this is the current working version
 # it's also possible to build customized BT
 worktools = b3.Priority([algSol, sc_tan, Simu_Eqn_Sol, sacSol, x2z2_Solver])
+
+
 
 subtree = b3.RepeatUntilSuccess(b3.Sequence([asgn, worktools]), 6)
 solveRoutine = b3.Sequence([sub_trans, subtree,  updateL, compDetect])
