+++ ikbtbasics/ik_classes.py
@@ -78,9 +78,8 @@ def kinematics_pickle(rname, dh, constants, pvals, vv, unks, test):
     if(os.path.isfile(name)):
         with open(name, 'rb') as pick:
             print '\Trying to read pre-computed forward kinematics from '+name
-            [m, R, unknowns]  = pickle.load(pick)
+            [m, R, unks]  = pickle.load(pick)
             print 'Successfully read pre-computed forward kinematics'
-            print 'pickle contained ', len(unknowns), ' unknowns'
     else:
         #print 'WRONG - quitting, error: ',sys.exc_info()[0]
         #sys.exit
@@ -105,9 +104,8 @@ def kinematics_pickle(rname, dh, constants, pvals, vv, unks, test):
         print ' Storing kinematics pickle for '+rname + '('+name+')'
         with open(name,'wb') as pf:
             pickle.dump( [m, R, unks], pf, protocol=pprotocol)
-        unknowns = unks    # be sure to return updated unknown list (including SOAs)
 
-    return [m,R,unknowns]
+    return [m,R,unks]
 
 
 def check_the_pickle(dh1, dh2):   # check that two mechanisms have identical DH params
@@ -153,9 +151,6 @@ class Robot:
         self.params = []  # constant dh params such as l_4 etc.
         self.solution_nodes = []  # first one is the root, by solve order
         self.variables_symbols = []
-        #
-        #   "notations" means specifically labeled solution variables such as
-        #          th_1s2  (theta-1, solution 2)
         self.notation_graph = set() #solution nodes notation graph
         self.notation_collections = [] #solution notations divided into subgroups
 
@@ -281,20 +276,28 @@ class Robot:
         print 'Starting sum-of-angles scan. Please be patient'
         unkn_sums_sym = set() #keep track of joint variable symbols
 
+        thx = sp.Wild('thx') # a theta_x term
+        thy = sp.Wild('thy') # a theta_x term
+        sgn = sp.Wild('sgn') # 1 or -1
+
+        aw = sp.Wild('aw')
+        bw = sp.Wild('bw')
+        cw = sp.Wild('cw')
+        s1 = sp.Wild('s1')
+        s2 = sp.Wild('s2')
+
         #k = equation number
         #i = row, j=col
         nits = len(self.mequation_list) * 3 * 4 # total number of equations
         barlen = nits/2
         it_number = 0
-        for k in range(0,len(self.mequation_list)):  # contains duplicates
+        for k in range(0,len(self.mequation_list)):
             Meq = self.mequation_list[k]  # get next matrix equation
             for i in [0,1,2]:   # only first three rows are interesting
                 for j in [0,1,2,3]:
                     it_number += 1
                     #print ' .. '
                     prog_bar(it_number, nits, barlen, 'Sum of Angles')
-                    #print 'kij: ', k,i,j
-                    
                     #print 'Sum of Angles: eqn,row,col: ', k,i,j
                     # simplify with lasting effect (note: try sp.trigsimp() for faster????)
                     Meq.Ts[i,j] = sp.simplify(Meq.Ts[i,j])  # simplify should catch c1s2+s1c2 etc. (RHS)
@@ -303,22 +306,77 @@ class Robot:
                     lhs = Meq.Td[i,j]
                     rhs = Meq.Ts[i,j]
 
-                        
-                    # simplify LHS
-                    lhs, newj, newe = sum_of_angles_sub(lhs, variables)
-                    if newj:
-                        variables.append(newj)
-                    if newe:
-                        self.kequation_aux_list.append(newe)
-                    # simplify RHS
-                    rhs, newj, newe= sum_of_angles_sub(rhs, variables)
-                    if newj:
-                        variables.append(newj)
-                    if newe:
-                        self.kequation_aux_list.append(newe)
-                    
-                    Meq.Td[i,j] = lhs 
-                    Meq.Ts[i,j] = rhs
+
+                    for expr in [lhs, rhs]:
+                        found2 = found3 = False
+                        matches = expr.find(sp.sin(aw+bw+cw)) | expr.find(sp.cos(aw+bw+cw))
+
+                        #print '- -  - - -'
+                        #print expr
+                        #print matches
+
+                        for m in matches: # analyze each match
+                            d  = m.match(sp.cos(aw + bw + cw))
+                            d1 = m.match(sp.sin(aw + bw + cw))
+                            if d != None and d1 != None:
+                                d.update(d1)
+                            if d == None and d1 != None:
+                                d = d1
+                            #  To find sum-of-angles arguments,
+                            # count number of non-zero elements among aw ... cw
+                            nzer = 0
+                            varlist = []
+                            for k1 in d.keys():
+                                if d[k1] == 0:
+                                    nzer += 1
+                                else:
+                                    varlist.append(d[k1])
+                            #print 'varlist: ', varlist
+                            if len(varlist) == 2:
+                                found2 = True
+                            if len(varlist) == 3:
+                                found3 = True
+
+                            if(found2 or found3):  # we've got a SOA!
+
+                                # generate index of the SOA variable
+                                nil = [] #new index list = 'nil'
+                                for v in varlist: # build the new subscript
+                                    nil.append( str(get_variable_index(variables, v)) )
+                                nil.sort()  # get consistent order of indices
+                                ni = ''
+                                for c in nil:  # make into a string
+                                    ni += c
+
+                                #print 'New index: '+ni
+                                exists = False
+                                # has this SOA been found before?  Did we already make it?
+                                for v in variables:
+                                    if v.n == int(ni):   # i.e. if th_23 is aready defined
+                                        th_subval = v
+                                        exists = True
+                                if not exists:
+                                    th_new = sp.var('th_'+ni) # create iff doesn't yet exists
+                                    th_subval = th_new
+                                    print "found new 'joint' (sumofangle) variable: (k=",k,") ", th_new
+                                    #  try moving soa equation to Tm.auxeqns
+                                    #unkn_sums_sym.add(th_new) #add into the joint variable set
+                                    newjoint = unknown(th_new)
+                                    newjoint.n = int(ni)  # generate e.g. 234 = 10*2 + 34
+                                    newjoint.solved = False  # just to be clear for count_unknowns
+                                    variables.append(newjoint) #add it to unknowns list
+                                    tmpeqn = kc.kequation(th_new, d[aw] + d[bw] + d[cw])
+                                    print 'sumofanglesT: appending new equation:', tmpeqn
+                                    self.kequation_aux_list.append(tmpeqn)
+
+                                # substitute new variable into the kinematic equations
+                                self.mequation_list[k].Td[i,j] = Meq.Td[i,j].subs(d[aw] + d[bw] + d[cw], th_subval)
+                                self.mequation_list[k].Ts[i,j] = Meq.Ts[i,j].subs(d[aw] + d[bw] + d[cw], th_subval)
+                                #print 'sum of angles (ik_classes): NEW Eqns (k,i,j)',k,i,j
+                                #print self.mequation_list[k].Td[i,j]
+                                #print ' = '
+                                #print self.mequation_list[k].Ts[i,j]
+                                #print '========'
 
         prog_bar(-1,100,100, '')  # clear the progress bar
 
@@ -326,104 +384,12 @@ class Robot:
         print 'Completed sum-of-angles scan.'
 
 
-##################
-# 
-#   substitute th_23 for th_2+th_3 etc.
-# (april: separate out for easier testing)
 
-def sum_of_angles_sub(expr, variables):   
-    thx = sp.Wild('thx') # a theta_x term
-    thy = sp.Wild('thy') # a theta_x term
-    sgn = sp.Wild('sgn') # 1 or -1
-    aw = sp.Wild('aw')
-    bw = sp.Wild('bw')
-    cw = sp.Wild('cw')
-    s1 = sp.Wild('s1')
-    s2 = sp.Wild('s2') 
-
-    newjoint = None
-    tmpeqn = None
-    found2 = found3 = False
-    matches = expr.find(sp.sin(aw+bw+cw)) | expr.find(sp.cos(aw+bw+cw))
-    #print '- -  - - -'
-    #print expr
-    #print matches
-
-    for m in matches: # analyze each match
-        d  = m.match(sp.cos(aw + bw + cw))
-        d1 = m.match(sp.sin(aw + bw + cw))
-        if d != None and d1 != None:
-            d.update(d1)
-        if d == None and d1 != None:
-            d = d1
-        #  To find sum-of-angles arguments,
-        # count number of non-zero elements among aw ... cw
-        nzer = 0
-        varlist = []
-        for k1 in d.keys():
-            if d[k1] == 0:
-                nzer += 1
-            else:
-                varlist.append(d[k1])
-        #print 'varlist: ', varlist
-        if len(varlist) == 2:
-            found2 = True
-        if len(varlist) == 3:
-            found3 = True
-
-        newjoint = None
-        tmpeqn = None
-        if(found2 or found3):  # we've got a SOA!
-
-            # generate index of the SOA variable
-            nil = [] #new index list = 'nil'
-            for v in varlist: # build the new subscript
-                nil.append( str(get_variable_index(variables, v)) )
-            nil.sort()  # get consistent order of indices
-            ni = ''
-            for c in nil:  # make into a string
-                ni += c
-
-            #print 'New index: '+ni
-            vexists = False
-            # has this SOA been found before?  Did we already make it?
-            for v in variables:
-                if v.n == int(ni):   # i.e. if th_23 is aready defined
-                    th_subval = v
-                    vexists = True
-            newjoint = None
-            tmpeqn = None
-            th_new = sp.var('th_'+ni) # create iff doesn't yet exist
-            th_subval = th_new
-            if not vexists:
-                print ":  found new 'joint' (sumofangle) variable: ", th_new
-                #  try moving soa equation to Tm.auxeqns
-                #unkn_sums_sym.add(th_new) #add into the joint variable set
-                newjoint = kc.unknown(th_new)
-                newjoint.n = int(ni)  # generate e.g. 234 = 10*2 + 34
-                newjoint.solved = False  # just to be clear for count_unknowns
-                variables.append(newjoint) #add it to unknowns list
-                tmpeqn = kc.kequation(th_new, d[aw] + d[bw] + d[cw])
-                print 'sum_of_angles_sub: created new equation:', tmpeqn
-                
-                #self.kequation_aux_list.append(tmpeqn)
-
-            # substitute new variable into the kinematic equations  ((WHY twice??))
-            #self.mequation_list[k].Td[i,j] = Meq.Td[i,j].subs(d[aw] + d[bw] + d[cw], th_subval)
-            #self.mequation_list[k].Ts[i,j] = Meq.Ts[i,j].subs(d[aw] + d[bw] + d[cw], th_subval)
-            expr = expr.subs(d[aw] + d[bw] + d[cw], th_subval)
-            #print 'sum of angles (ik_classes): NEW Eqns (k,i,j)',k,i,j
-            #print self.mequation_list[k].Td[i,j]
-            #print ' = '
-            #print self.mequation_list[k].Ts[i,j]
-            #print '========'
- 
-    return (expr,newjoint, tmpeqn)
 
 def get_variable_index(vars, symb):
     for v in vars:
         if v.n == 0:
-            print 'get_variable_index()/ik_classes: at least one index is not initialized for joint variables (or is 0!)'
+            print 'get_variable_index()/ik_classes: at least one index is not initialized for joint variables'
             quit()
         found = False
         #print 'get_variable_index: v[i], symb, v[i].n ',str(v.symbol),str(symb), v.n
@@ -441,7 +407,131 @@ def get_variable_index(vars, symb):
 
 
 # class kequation()       now moved to kin_cl.py
-# class unknown(object)   now moved to kin_cl.py
+
+class unknown(object):
+    def __init__(self,u=sp.var('x'), mat_eqn=None):
+        self.symbol = u
+        self.n = 0           # index of the unk in the serial chain (1-6) 0=unset
+                        #  NEW: self.n can be 23 e.g. for (th2+th3) or 234 for (th2+th3+th4)
+        self.eqnlist = []    # list of kequations containing this UNK
+        self.readytosolve = False
+        self.eqntosolve = None #this has to be NONE, otherwise the None judgement in tan_solver wouldn't work
+
+        self.secondeqn  = None
+
+        self.solvemethod = "*None*"
+        self.solved = False
+        self.solveorder = 0
+        self.usedfortransform = False   # if solved, has this been used for transform yet?
+        self.solutions = []   # list of solutions, store final solutions
+        self.argument = sp.var('a')*sp.var('b')  # argument to arcin() for example (used for generating checking code output)
+
+        # for nodes ranking
+        self.sincos_solutions = [] # solutions from sin or cos
+        self.sincos_eqnlist = []
+        self.solvable_sincos = False
+        self.tan_solutions = [] #solutions from tangent solvers
+        self.tan_eqnlist = []
+        self.solvable_tan = False
+        # end: nodes ranking
+        self.nsolutions = 0   # number of solutions (== len(self.solutions))
+        self.assumption = [] #assumputions about the solutions
+        #self.nodelist = []   # list of solution tree nodes for this variable
+        if mat_eqn != None:     #  list of kequation scontaining this unknown
+            self.scan(mat_eqn)
+
+    def __eq__(self, other): #equal judgement, also hashing in python 3.x
+        if other != None:
+            return self.symbol == other.symbol
+        return False
+
+    def __hash__(self): #hash function "inherits" from symbol
+        return self.symbol.__hash__()
+
+    def __repr__(self): # string representation
+        return self.symbol.__repr__()
+
+    # class unknown:
+    def set_solved(self, R, unknowns):           # indicate that a this variable has been solved
+                                                 #  and update the solution tree
+        self.solved = True
+        self.readytosolve = False
+        print '\n\n'
+        print 'set_solved: ', self.symbol, '      by: ', self.solvemethod
+        #print '            ', self.eqntosolve
+        fs = 'set_solved: solutions empty '
+        assert (len(self.solutions) >= 1), fs
+        assert (self.nsolutions > 0), fs
+        print '            ', self.symbol, '=', self.solutions[0], '\n\n'
+        #print 'Robot instance.name: ', R.name      # shouldn't change!!
+        #########################################
+        #
+        #     Update Solution Tree
+        #
+        R.solveN += 1                 # increment solution level counter
+        self.solveorder = R.solveN    # first solution starts with 1 (0 is the root)
+
+        curr_node = None
+
+        found = False
+        for sol_node in R.solution_nodes:   # make sure there is a node for this var
+            if sol_node.symbol == self.symbol:
+                found = True
+
+        if not found:
+            R.solution_nodes.append(Node(self))
+            R.variables_symbols.append(self.symbol)
+
+        # for new solution graph
+        for sol_node in R.solution_nodes:
+            if sol_node.symbol == self.symbol:
+                curr_node = sol_node
+
+        print ' -  - - - - '
+        print R.solution_nodes
+        print 'Trying to find: ',  self.symbol
+        print ' - - - - - '
+        assert(curr_node is not None), ' Trouble finding solution tree node'
+
+        curr_node.solveorder = R.solveN
+        curr_node.solvemethod = self.solvemethod
+        curr_node.argument = self.argument
+        curr_node.nsolutions = self.nsolutions
+        curr_node.solutions = self.solutions
+        curr_node.assumption = self.assumption
+
+        # set the equations
+        curr_node.eqnlist.append(self.eqntosolve)
+
+        if self.secondeqn is not None:
+            curr_node.eqnlist.append(self.secondeqn)
+
+        #set solutions
+
+        curr_node.detect_parent(R)
+        curr_node.generate_notation(R)
+        #curr_node.generate_solutions(R)
+        print 'finish set_solved', self.symbol
+
+    def scan(self,MatEqn):        # find list of kequations containing this UNK
+        self.eqnlist = []   # reset eqn list
+        rng = [0,1,2,3]
+        for i in rng:
+            for j in rng:
+                eqn = MatEqn.Ts[i,j]
+                if (eqn != 0):
+                    if eqn.has(self.symbol):
+                        self.eqnlist.append(kc.kequation(MatEqn.Td[i,j],eqn))
+
+                eqn = MatEqn.Td[i,j]
+                if (eqn != 0):
+                    if eqn.has(self.symbol):
+                        #print "Equation [", eqn.string, "] has ", self.symbol
+                        self.eqnlist.append(kc.kequation(MatEqn.Ts[i,j],eqn))
+
+
+        self.eqnlist = erank(self.eqnlist)  # sort them in place
+
 
 # matrix_equation class moved to kin_cl.py
 
@@ -666,8 +756,6 @@ def erank(list_L):    # rearrange list of eqns by length
 
 
 
-#############    main       test the library  #########################
-#
 if __name__ == "__main__":   # tester code for the classes in this file
 
     #j1 = joint_var(th_12)
@@ -718,8 +806,8 @@ if __name__ == "__main__":   # tester code for the classes in this file
 
      ####Test unknown class
 
-    ua = kc.unknown(a)
-    ub = kc.unknown(b)
+    ua = unknown(a)
+    ub = unknown(b)
 
     print "\n\nTesting unknown(symbol) (one-arg form)"
     print "Unknown a: ",   ua.symbol
@@ -783,8 +871,8 @@ if __name__ == "__main__":   # tester code for the classes in this file
 
 
     # unknown class hash function testing
-    tmpv1 = kc.unknown(th_1)
-    tmpv2 = kc.unknown(th_1)
+    tmpv1 = unknown(th_1)
+    tmpv2 = unknown(th_1)
     print 'a.symbol: ', tmpv1
     print 'b.symbol: ', tmpv2
     c = set()
@@ -803,92 +891,8 @@ if __name__ == "__main__":   # tester code for the classes in this file
     #   Robot class is tested in updateL.py
     print 'Sum of Angles testing'
     #
-    
-    
-    ############################################################3
-    #
-    #  basic sum of angles testing:
-    s = 'Basic Sum of Angles Testing'
-    print '\n\n ' + s + '\n\n'
-    
-    sp.var('a1 a2 a3')   # note subscripts required
-    
-    vars01 = [kc.unknown(a1), kc.unknown(a2), kc.unknown(a3)]
-    
-    i=1
-    for v in vars01:   # variables need an index to place them in mechanism order
-                       #   must start with 1 for DH/Craig compatibility 
-        v.n = i
-        i+=1 
-   
-    term1 = sp.sin(a1+a2)
-    term2 = sp.sin(a1+a2+a3)
-    term1a, newj, newe = sum_of_angles_sub(term1, vars01)
-    fs = ' new equation not correctly established'
-    assert str(newj.symbol) == 'th_12', fs
-    print 'new eqn', newe
-    term2a, newj, newe = sum_of_angles_sub(term2, vars01)
-    assert str(newj.symbol) == 'th_123', fs
-    print 'new eqn', newe
-    
-    #print 'Initial Test: '
-    #print term1, ' --> ', term1a
-    #print term2, ' --> ', term2a
-    
-    fs = 'Sum of angles: basic test fail'
-    assert term1a == sp.sin(th_12), fs
-    assert term2a == sp.sin(th_123), fs
-    
 
-    #####################################################################
-    #
-    #   Problem-specific SOA test
-    #
-
-    s = 'Problem Specific Sum of Angles Testing'
-    print '\n\n ' + s + '\n\n'
-    
-    sp.var('Px Py Pz')
-    unks01 =  [kc.unknown(th_1), kc.unknown(th_2), kc.unknown(th_3), kc.unknown(th_4), kc.unknown(th_5), kc.unknown(th_6)]
-    i=1
-    for v in unks01:   # variables need an index to place them in mechanism order
-                       #   must start with 1 for DH/Craig compatibility 
-        v.n = i
-        i+=1 
-   
-
-    eqnterm = Px*sp.sin(th_2 + th_3 + th_4)*sp.cos(th_1) + Py*sp.sin(th_1)*sp.sin(th_2 + th_3 + th_4) + Pz*sp.cos(th_2 + th_3 + th_4) - a_2*sp.sin(th_3 + th_4) - a_3*sp.sin(th_4) - d_1*sp.cos(th_2 + th_3 + th_4)
-    
-    term2, newj, newe = sum_of_angles_sub(eqnterm, unks01)
-    if newj:
-        unks01.append(newj)
-    if newe:
-        print ' NEW Equation: ', newe
-        
-    print '---'
-    print eqnterm
-    print '---'
-    print term2
-    print '---'
-    print 'new unknown list: ', unks01
-    
-    fs = 'FAIL problem specific SOA tests'
-    assert str(term2) == 'Px*sin(th_234)*cos(th_1) + Py*sin(th_1)*sin(th_234) + Pz*cos(th_234) - a_2*sin(th_34) - a_3*sin(th_4) - d_1*cos(th_234)', fs
-    assert kc.unknown(th_234) in unks01, fs
-    assert kc.unknown(th_34) in unks01, fs
- 
- 
- 
-    #####################################################
-    #
-    #  advanced SOA testing 
-    #
     # Generate some SUM of Angles Kin eqns
-    
-    s = 'Advanced Sum of Angles Testing: DH parameters/UR5'
-    print '\n\n ' + s + '\n\n'
-    
-    
     dh = sp.Matrix([
             [    0    ,    0 ,  d_1 ,     th_1  ],  # based on UR5 but simplified
             [-sp.pi/2 ,    0 ,   0  ,     th_2  ],
@@ -899,7 +903,7 @@ if __name__ == "__main__":   # tester code for the classes in this file
             ])
     vv = [1,1,1,1,1,1]
 
-    variables =  [kc.unknown(th_1), kc.unknown(th_2), kc.unknown(th_3), kc.unknown(th_4), kc.unknown(th_5), kc.unknown(th_6)]
+    variables =  [unknown(th_1), unknown(th_2), unknown(th_3), unknown(th_4), unknown(th_5), unknown(th_6)]
     i=1
     for v in variables:
         v.n = i   # set the variable index (joint number)
@@ -911,19 +915,15 @@ if __name__ == "__main__":   # tester code for the classes in this file
     robot = 'SOA Test Robot'
 
     testing = False  # not using this now
-    [m, R, tmpvars] = kinematics_pickle(robot, dh, params, pvals, vv, variables, testing)
+    [m, R, unknowns] = kinematics_pickle(robot, dh, params, pvals, vv, variables, testing)
     print 'GOT HERE: robot name: ', R.name
 
-    variables = tmpvars
     R.name = robot
     R.params = params
 
     ##   check the pickle in case DH params were changed
     check_the_pickle(m.DH, dh)   # check that two mechanisms have identical DH params
 
-    print '\n the variables: ', variables
-    assert len(variables) == 9, 'wrong number of variables'
-    print '\n\n'
 
     # Below replaced by pickle code above
 
@@ -934,32 +934,19 @@ if __name__ == "__main__":   # tester code for the classes in this file
     #print "Completed Forward Kinematics"
     ##print 'Starting Sum of Angles scan (slow!)'
     # set up Robot Object instance
-    #R = Robot(m, 'SOA 2,3 TEST Robot')              # set up IK structs et
-    #R.scan_for_equations(variables)       # generate equation lists
-    
-    
-    
-    ## below is commented out for testing and devel of sum_of_angles_transform
-    
-    # need to re-do for debugging if test fails. 
-    #R.sum_of_angles_transform(variables)  # find sum of angles
+    #R = Robot(m, 'SOA 2,3 TEST Robot')              # set up IK structs etc
 
 
 
-    ####   Temp code:  manually review all the FK matrix equations(!)
-    eqn = 0
-    row = 2
-    col = 3
-    print 'Equations: '
-    enbr = 0
-    for eqn in R.mequation_list:
-        print enbr, ':  ', eqn.Td[row,col], '    =    ', eqn.Ts[row,col]
-        enbr += 1
-    print '\n\n\n'
-    ####
+    R.scan_for_equations(variables)       # generate equation lists
+    # below is commented out for testing and devel of sum_of_angles_transform
+    R.sum_of_angles_transform(variables)  # find sum of angles
+    fs = 'ik_classes: sum_of_angles_transform FAILS 3-way sum'
 
+    print '\n the variables: ', variables
+    assert len(variables) == 9, 'wrong number of variables'
+    print '\n\n'
 
-    
     #  Now we test some expected RHS results
     sp.var('A1 A2') # two possible answers
 
@@ -971,19 +958,7 @@ if __name__ == "__main__":   # tester code for the classes in this file
     # expected correct answer:
     A2 = -(sp.sin(th_1)*sp.sin(th_5) - sp.cos(th_1)*sp.cos(th_234)*sp.cos(th_5))*sp.cos(th_6) + sp.sin(th_234)*sp.sin(th_6)*sp.cos(th_1)
 
-
-
-    print '\n\n\n'
-    print '--- (round 1)'
-    print A1          ##   failing to substitute 3-way SOA right now
-    print '---'
-    print A2
-    print '---'
-    print sp.simplify(A1-A2)
-    print '\n\n\n'
-    
-    fs = 'ik_classes: sum_of_angles_transform FAILS 3-way sum'
-    assert sp.simplify(A1-A2)== 0, fs+' 001'
+    assert sp.simplify(A1-A2)== 0, fs
 
 
     # Assertion 2
@@ -997,14 +972,11 @@ if __name__ == "__main__":   # tester code for the classes in this file
     A2 =  Px*sp.sin(th_234)*sp.cos(th_1) + Py*sp.sin(th_1)*sp.sin(th_234) + Pz*sp.cos(th_234) - a_2*sp.sin(th_34) - a_3*sp.sin(th_4) - d_1*sp.cos(th_234)
 
     print '\n\n\n'
-    print '--- (round 2)'
-    print A1          ##   failing to substitute 3-way SOA right now
-    print '---'
+    print A1
     print A2
-    print '---'
     print sp.simplify(A1-A2)
     print '\n\n\n'
-    assert sp.simplify(A1-A2)==0, fs+' 002'
+    assert sp.simplify(A1-A2)==0, fs
 
 
     print '\n\n\n        ik_classes   PASSES all tests \n\n'
