+++ ikbtbasics/kin_cl.py
@@ -84,134 +84,8 @@ class kequation:
         tmp = tmp.replace(r'th_', r'\theta_')     # change to greek theta
         tmp = re.sub(r'_(\d+)',  r'_{\1}', tmp)   # get all digits of subscript into {}
         tmp = re.sub(r'atan_\{2\}','atan2' , tmp)  # correct atan2 formatting
+
         return tmp
-    
-    
-class unknown(object):
-    def __init__(self,u=sp.var('x'), mat_eqn=None):
-        self.symbol = u
-        self.n = 0           # index of the unk in the serial chain (1-6) 0=unset
-                        #  NEW: self.n can be 23 e.g. for (th2+th3) or 234 for (th2+th3+th4)
-        self.eqnlist = []    # list of kequations containing this UNK
-        self.readytosolve = False
-        self.eqntosolve = None #this has to be NONE, otherwise the None judgement in tan_solver wouldn't work
-
-        self.secondeqn  = None
-
-        self.solvemethod = "*None*"
-        self.solved = False
-        self.solveorder = 0
-        self.usedfortransform = False   # if solved, has this been used for transform yet?
-        self.solutions = []   # list of solutions, store final solutions
-        self.argument = sp.var('a')*sp.var('b')  # argument to arcin() for example (used for generating checking code output)
-
-        # for nodes ranking
-        self.sincos_solutions = [] # solutions from sin or cos
-        self.sincos_eqnlist = []
-        self.solvable_sincos = False
-        self.tan_solutions = [] #solutions from tangent solvers
-        self.tan_eqnlist = []
-        self.solvable_tan = False
-        # end: nodes ranking
-        self.nsolutions = 0   # number of solutions (== len(self.solutions))
-        self.assumption = [] #assumputions about the solutions
-        #self.nodelist = []   # list of solution tree nodes for this variable
-        if mat_eqn != None:     #  list of kequation scontaining this unknown
-            self.scan(mat_eqn)
-
-    def __eq__(self, other): #equal judgement, also hashing in python 3.x
-        if other != None:
-            return self.symbol == other.symbol
-        return False
-
-    def __hash__(self): #hash function "inherits" from symbol
-        return self.symbol.__hash__()
-
-    def __repr__(self): # string representation
-        return self.symbol.__repr__()
-
-    # class unknown:
-    def set_solved(self, R, unknowns):           # indicate that a this variable has been solved
-                                                 #  and update the solution tree
-        self.solved = True
-        self.readytosolve = False
-        print '\n\n'
-        print 'set_solved: ', self.symbol, '      by: ', self.solvemethod
-        #print '            ', self.eqntosolve
-        fs = 'set_solved: solutions empty '
-        assert (len(self.solutions) >= 1), fs
-        assert (self.nsolutions > 0), fs
-        print '            ', self.symbol, '=', self.solutions[0], '\n\n'
-        #print 'Robot instance.name: ', R.name      # shouldn't change!!
-        #########################################
-        #
-        #     Update Solution Tree
-        #
-        R.solveN += 1                 # increment solution level counter
-        self.solveorder = R.solveN    # first solution starts with 1 (0 is the root)
-
-        curr_node = None
-
-        found = False
-        for sol_node in R.solution_nodes:   # make sure there is a node for this var
-            if sol_node.symbol == self.symbol:
-                found = True
-
-        if not found:
-            R.solution_nodes.append(Node(self))
-            R.variables_symbols.append(self.symbol)
-
-        # for new solution graph
-        for sol_node in R.solution_nodes:
-            if sol_node.symbol == self.symbol:
-                curr_node = sol_node
-
-        #print ' -  - - - - '
-        #print R.solution_nodes
-        #print 'Trying to solution tree node for: ',  self.symbol
-        #print ' - - - - - '
-        assert(curr_node is not None), ' Trouble finding solution tree node'
-
-        curr_node.solveorder = R.solveN
-        curr_node.solvemethod = self.solvemethod
-        curr_node.argument = self.argument
-        curr_node.nsolutions = self.nsolutions
-        curr_node.solutions = self.solutions
-        curr_node.assumption = self.assumption
-
-        # set the equations
-        curr_node.eqnlist.append(self.eqntosolve)
-
-        if self.secondeqn is not None:
-            curr_node.eqnlist.append(self.secondeqn)
-
-        #set solutions
-
-        curr_node.detect_parent(R)
-        curr_node.generate_notation(R)
-        #curr_node.generate_solutions(R)
-        #print 'finish set_solved', self.symbol
-        print '\n\n'
-
-    def scan(self,MatEqn):        # find list of kequations containing this UNK
-        self.eqnlist = []   # reset eqn list
-        rng = [0,1,2,3]
-        for i in rng:
-            for j in rng:
-                eqn = MatEqn.Ts[i,j]
-                if (eqn != 0):
-                    if eqn.has(self.symbol):
-                        self.eqnlist.append(kc.kequation(MatEqn.Td[i,j],eqn))
-
-                eqn = MatEqn.Td[i,j]
-                if (eqn != 0):
-                    if eqn.has(self.symbol):
-                        #print "Equation [", eqn.string, "] has ", self.symbol
-                        self.eqnlist.append(kc.kequation(MatEqn.Ts[i,j],eqn))
-
-
-        self.eqnlist = erank(self.eqnlist)  # sort them in place
-
 
 class matrix_equation:
     def __init__(self, Td=sp.zeros(4), Ts=sp.zeros(4)):
