+++ ikbtleaves/algebra_solver.py
@@ -29,7 +29,6 @@ from ikbtbasics.ik_classes import *     # special classes for Inverse kinematics
 
 import b3 as b3          # behavior trees
 from assigner_leaf  import *
-import comp_detect as cpd
 
 
  
@@ -46,35 +45,23 @@ class test_algebra_id(b3.Action):    # tester for your ID
         
         Ts[0,1] =  th_2 + l_1*l_2
         Ts[0,2] =  d_1*l_3 + l_1
-        #Ts[0,3] = sp.sin(th_1)*l2 + l_1
-        
-        Ts[1,1] =  th_5 + th_2*l_1
-        Ts[1,2] =  0
-
 
         Ts[2,1] =  th_2 * th_3 + l_1
         Ts[2,2] =  sp.sin(th_3)
 
-        Ts[3,1] = sp.sin(th_1 + th_2)
-        Ts[3,2] = sp.sin(th_1 + th_2 + th_3)
-        
+        Ts[1,1] =  th_5 + th_2*l_1
+        Ts[1,2] =  0
 
         testm = matrix_equation(Td,Ts)
-        sp.var('th_12 th_123 B')  # B is "known"
 
         ud1  = unknown(d_1)
-        u12 = unknown(th_12)
-        u123 = unknown(th_123)
-        uth1 = unknown(th_1)
         uth2 = unknown(th_2)
         uth3 = unknown(th_3)
         uth4 = unknown(th_4)
         uth5 = unknown(th_5)
-        variables = [ud1,  uth1, uth2, uth3, uth4, uth5,u123]
+        variables = [ud1, uth2, uth3, uth4, uth5]
 
         [L1, L2] = R.scan_Mequation(testm, variables)  # lists of 1unk and 2unk equations
-        
-        L1.append(kequation(th_123, th_1+th_2+B))
         
         # fix!!
         R.generate_solution_nodes(variables) # generate solution nodes
@@ -93,10 +80,6 @@ class algebra_id(b3.Action):    # action leaf for
         unknowns = tick.blackboard.get('unknowns')   # the current list of unknowns
         
         one_unk = tick.blackboard.get('eqns_1u')
-        if(len(one_unk)==0):   # algebra-solver can only work on one-unknown equations
-            print 'algebra_id: there are no one-unknown equations'
-            return b3.FAILURE
-        
         two_unk = tick.blackboard.get('eqns_2u')
         R = tick.blackboard.get('Robot')
             
@@ -115,7 +98,7 @@ class algebra_id(b3.Action):    # action leaf for
                     if(self.BHdebug):
                         print 'algebra ID: Looking for unknown: ', u.symbol, ' in equation: ', 
                         print e ,
-                        print "  - ", count_unknowns(unknowns, e.RHS), " unknown in RHS"
+                        print "  - ", count_unknowns(unknowns, e.RHS), " unknown"
                     if (e.RHS.has(sp.sin(u.symbol)) or e.RHS.has(sp.cos(u.symbol)) or\
                         e.LHS.has(sp.sin(u.symbol)) or e.LHS.has(sp.cos(u.symbol))):
                         continue   # this shouldbe caught by another ID
@@ -173,7 +156,7 @@ class algebra_solve(b3.Action):    # Solve asincos equation pairs
 #  Test code:
 class TestSolver002(unittest.TestCase):
     def setUp(self):
-        self.DB = True  # debug flag
+        self.DB = False  # debug flag
         print '\n\n===============  Test algebra Solver  ====================='
         return
     
@@ -184,7 +167,7 @@ class TestSolver002(unittest.TestCase):
         algebra_tester = b3.BehaviorTree()
         bb = b3.Blackboard()  
         bb.set('Robot', Robot())
-        setup = test_algebra_id()  # see top of this file
+        setup = test_algebra_id()
         aid   = algebra_id()
         aid.Name = 'Algebra ID'
         aid.BHdebug = self.DB
@@ -192,14 +175,10 @@ class TestSolver002(unittest.TestCase):
         ais.Name = 'Algebra Solver'
         ais.BHdebug = self.DB
         
-        compdet = cpd.comp_det()
-        compdet.Name = 'Completion Checker'
-        compdet.BHdebug = self.DB
-        
         asgn = assigner()
-        subtree = b3.Sequence([asgn, aid, ais,compdet])
+        subtree = b3.Sequence([asgn, aid, ais])
         
-        test = b3.Sequence([setup, b3.Repeater(subtree, max_loop = 5)])          
+        test = b3.Sequence([setup, b3.Repeater(subtree, max_loop = 6)])          
         algebra_tester.root = test
         
         # Run the testing BT 
@@ -229,7 +208,6 @@ class TestSolver002(unittest.TestCase):
                 self.assertFalse(u.solved, fs)
 
         self.assertTrue(ntests == 3, ' Algebra solver:  assertion count error --- FAIL')
-        print 'Algebra solver PASSED ', ntests, ' assertions.'
 ##  write tester code which runs if this file is run directly instead
 ##    of "imported".
