+++ ikbtleaves/comp_detect.py
@@ -41,24 +41,14 @@ class comp_det(b3.Action):
         unks = tick.blackboard.get('unknowns')
         Tm = tick.blackboard.get('Tm')
         R  = tick.blackboard.get('Robot')
-        
-        SOA_DEBUG = False  # just turn this on for debugging
-        if SOA_DEBUG:
-            L1 = tick.blackboard.get('eqns_1u')
-            L2 = tick.blackboard.get('eqns_2u')
-            L3 = tick.blackboard.get('eqns_3pu')
-            print '\n'
-            print 'L1: ', L1   ##  for debugging sum of angles
-            print 'L2: ', L2
-            print 'L3: ', L3
-            
         n = 0
         ns = 0
         for u in unks:
             n += 1
             if(u.solved):
                 ns += 1
-        print '\n\n'
+        #if(self.BHdebug):
+        print '\n\n\n\n'
         print '           Completion Detector: ', n, ' variables, ', ns, ' are solved.'
         print '             solved: ',
         for u in unks:
